// src/campaigns/email-campaigns/campaigns/domain/types.ts
export enum CampaignStatus {
  ACTIVE = 'active',
  DRAFT = 'draft',
  STOPPED = 'stopped',
  CANCELED = 'canceled',
}

export type ContentId = number;
export type TemplateId = number;

export enum CampaignHistoryType {
  ABSOLUTE = 'absolute',
  RELATIVE = 'relative',
}

export enum CampaignHandler {
  ABSOLUTE = 'sendAbsoluteCampaign',
  RELATIVE = 'sendRelativeCampaign',
}

export type CampaignWithSegments = {
  segments: Array<SegmentDomain>;
} & CampaignDocument;

export type CampaignReport = {
  _id: SchemaId;
  name: string;
  templates: Array<string>;
  status: string;
  segments: Array<string>;
  totalSent: number;
  totalDelivery: number;
  totalBounce: number;
  totalComplaints: number;
  customer: Pick<CustomerDocument, '_id' | 'firstName' | 'lastName'>;
} & CampaignHistoryWithSegments;

export type CampaignHistoryWithSegments = {
  campaign: CampaignDocument;
  segments: Array<SegmentDomain>;
} & CampaignHistory;
// ...
export type CampaignTotalsCount = {
  delivery: number;
  bounce: number;
  unsubscribed: number;
  total: number;
  [key: string]: number;
};

export type CampaignBucket = {
  key: string;
  doc_count: number;
};

export type CampaignsAggregationBuckets = Array<CampaignBucket>;

export type CampaignsMetrics = {
  campaigns: number;
  history: PaginatorSchematicsInterface<CampaignReport>;
  totals: CampaignTotalsCount;
};

export type EmailMetrics = {
  data: {
    metrics: Array<{
      date: string;
      events: Array<{
        type: string;
        ref: string;
        createdAt: string;
        to: string;
        from: string;
      }>;
      meta: {
        [key: string]: number;
      };
    }>;
    metricsMeta: {
      Send: number;
      Open: number;
      Click: number;
      Bounce: number;
      Complaint: number;
      Reject: number;
      Delivery: number;
    };
    campaign: {
      id: Types.ObjectId;
      name: string;
      segments: ResponseSegmentsType[];
      allSegments: boolean;
      status: CampaignStatus;
      startDate: Date;
      allowWeekend: boolean;
      contentId: number;
      createdAt: string | boolean;
      updatedAt: string | boolean;
    };
  }[];
  total: {
    current: {
      Send: number;
      Open: number;
      Click: number;
      Bounce: number;
      Complaint: number;
      Delivery: number;
    };
    previous: {
      Send: number;
      Open: number;
      Click: number;
      Bounce: number;
      Complaint: number;
      Delivery: number;
    };
  };
  meta: {
    page: number;
    perPage: number;
    total: number;
  };
};